Made changes to the folder once again, changed html and css files, added modal to login and added code to the server.js and api-routes.js
