--- app/public/assets/css/map.css
@@ -1,58 +1,0 @@
-
-html, body
-{
-    height: 100%;
-    margin: 0;
-    padding: 0;
-}
-
-html {
-    overflow-x: hidden;
-    overflow-y: scroll;
-    position: relative;
-}
-
-nav {
-	position: absolute;
-	z-index: 3;
-}
-
-/* Styling for map*/
-#map {
-	height: 80%;
-    width: 100%;
-}
-
-#respond {
-    margin-top: 20px;
-    margin-bottom: 40px;
-}
-
-#respond input[type='text'],
-#respond input[type='email'],
-#respond textarea {
-    margin-bottom: 10px;
-    display: block;
-    width: 100%;
-    border: 1px solid rgba(0, 0, 0, 0.1);
-    -webkit-border-radius: 5px;
-    -moz-border-radius: 5px;
-    -o-border-radius: 5px;
-    -ms-border-radius: 5px;
-    -khtml-border-radius: 5px;
-    border-radius: 5px;
-    line-height: 1.4em;
-}
-
-footer {
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-    height: 60px;
-    line-height: 55px;  
-    /* color: #ffffff;
-    background-color: #2b5e64;
-    border-top: 6px solid #5f5d5d; */
-    text-align: center;
-} 
-
