+++ package.json
@@ -10,14 +10,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-<<<<<<< HEAD
     "body-parser": "^1.18.2",
     "express": "^4.16.3",
     "mysql2": "^1.5.2",
-=======
-    "express": "^4.16.3",
-    "express-handlebars": "^3.0.0",
->>>>>>> cf67675cbff37cb4b7df610b8ffe7bf765300fbf
     "save": "^2.3.2",
     "sequelize": "^4.37.1"
   }
