+++ public/assets/css/map.css
@@ -1,3 +1,4 @@
+
 html, body
 {
     height: 100%;
@@ -11,18 +12,24 @@ html {
     position: relative;
 }
 
-#mainPgContainer {
-    margin-top: 40px;
-    margin-bottom: 40px;
+nav {
+	position: absolute;
+	z-index: 3;
 }
 
-#missionContainer {
+#mapPgContainer {
+    margin-top: 10px;
+}
+
+#commentThreads{
+    margin-top: -120px;
     margin-bottom: 80px;
 }
 
-#recoContainer,  #loginContainer, #mapPgContainer {
-    margin-top: 20px;
-    margin-bottom: 80px;
+/* Styling for map*/
+#map {
+	height: 80%;
+    width: 100%;
 }
 
 footer {
@@ -36,3 +43,4 @@ footer {
     border-top: 6px solid #5f5d5d; */
     text-align: center;
 } 
+
