+++ public/assets/css/style.css
@@ -1,0 +1,39 @@
+html, body
+{
+    height: 100%;
+    margin: 0;
+    padding: 0;
+}
+
+html {
+    overflow-x: hidden;
+    overflow-y: scroll;
+    position: relative;
+}
+
+#mainPgContainer {
+    margin-top: 40px;
+    margin-bottom: 40px;
+}
+
+#missionContainer {
+    margin-bottom: 80px;
+}
+
+#recoContainer,  #loginContainer {
+    margin-top: 20px;
+    margin-bottom: 80px;
+}
+
+
+footer {
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+    height: 60px;
+    line-height: 55px;  
+    /* color: #ffffff;
+    background-color: #2b5e64;
+    border-top: 6px solid #5f5d5d; */
+    text-align: center;
+}
