+++ public/authentication.html
@@ -1,4 +1,4 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
 
 <head>
@@ -47,11 +47,11 @@
                 <form>
                     <div class="form-group">
                         <label for="exampleInputEmail1">Email Address</label>
-                        <input type="email" class="form-control" id="inputEmail" aria-describedby="emailHelp" placeholder="John@soandso.net">
+                        <input type="email" class="form-control" id="inputEmail" aria-describedby="emailHelp" placeholder="John@soandso.net" required>
                     </div>
                     <div class="form-group">
                         <label for="exampleInputPassword1">Password</label>
-                        <input type="password" class="form-control" id="inputPassword" placeholder="Enter Password">
+                        <input type="password" class="form-control" id="inputPassword" placeholder="Enter Password" required>
                     </div>
                     <button class="btn btn-primary" id="createAccount">Create Account</button>
                     <button class="btn btn-primary" id="login">Log in</button>
@@ -76,7 +76,7 @@
     crossorigin="anonymous"></script>
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
     crossorigin="anonymous"></script>
-<script src="public/assets/js/authentication.js"></script>
-<script src="public/assets/js/upvote.js"></script>
+
+<script src="assets/js/authentication.js"></script>
 
 </html>
